fix:페이지네이션 작동 오류 수정 및 변경된 api 반영
