+++ src/mocks/handlers/diary.js
@@ -9,41 +9,39 @@ const diaryHandlers = [
     (req, res, ctx) => {
       return res(
         ctx.status(200),
-        ctx.json({
-          data: [
-            {
-              id: 6,
-              title: '새 포스트6',
+        ctx.json([
+          {
+            id: 6,
+            title: '새 포스트6',
 
-              createAt: '2023-04-10T14:33:25.610691',
-            },
-            {
-              id: 5,
-              title: '새 포스트5',
-              createAt: '2023-04-10T14:33:23.179573',
-            },
-            {
-              id: 4,
-              title: '새 포스트4',
-              createAt: '2023-04-10T14:33:20.646264',
-            },
-            {
-              id: 3,
-              title: '새 포스트3',
-              createAt: '2023-04-10T14:33:18.170682',
-            },
-            {
-              id: 2,
-              title: '새 포스트2',
-              createAt: '2023-04-10T14:33:15.474825',
-            },
-            {
-              id: 1,
-              title: '새 포스트1',
-              createAt: '2023-04-10T14:33:12.378254',
-            },
-          ],
-        }),
+            createAt: '2023-04-10T14:33:25.610691',
+          },
+          {
+            id: 5,
+            title: '새 포스트5',
+            createAt: '2023-04-10T14:33:23.179573',
+          },
+          {
+            id: 4,
+            title: '새 포스트4',
+            createAt: '2023-04-10T14:33:20.646264',
+          },
+          {
+            id: 3,
+            title: '새 포스트3',
+            createAt: '2023-04-10T14:33:18.170682',
+          },
+          {
+            id: 2,
+            title: '새 포스트2',
+            createAt: '2023-04-10T14:33:15.474825',
+          },
+          {
+            id: 1,
+            title: '새 포스트1',
+            createAt: '2023-04-10T14:33:12.378254',
+          },
+        ]),
       );
     },
   ),
