--- src/pages/Mypage/Diary/Posts/List/Item/index.jsx
@@ -1,10 +1,0 @@
-const Item = ({ item }) => {
-  return (
-    <li>
-      <h1>{item.title}</h1>
-      <span>{item.createAt}</span>
-    </li>
-  );
-};
-
-export default Item;
