+++ src/pages/Mypage/Diary/index.jsx
@@ -5,7 +5,7 @@ import {
   requestDiaryNumByEmotion,
 } from '@/apis/request/diary';
 import useError from '@/hooks/useError';
-import Posts from './Posts';
+import Post from './Post';
 
 const LIMIT = 10;
 
@@ -13,12 +13,12 @@ const Diary = props => {
   const [filter, setFilter] = useState('best');
   const [list, setList] = useState([]);
   const [diaryNumber, setDairyNumber] = useState({});
-  const [currentPage, setCurrentPage] = useState(1);
+  const [currentPage, setCurrentPage] = useState(0);
   const handleError = useError();
 
   useEffect(() => {
     requestDiaryByEmotion({ mood: 'best', page: 0, size: LIMIT })
-      .then(({ data }) => {
+      .then(data => {
         setList(data);
       })
       .catch(error => {
@@ -36,7 +36,7 @@ const Diary = props => {
 
   useEffect(() => {
     requestDiaryByEmotion({ mood: filter, page: currentPage, size: LIMIT })
-      .then(({ data }) => {
+      .then(data => {
         setList(data);
       })
       .catch(error => {
@@ -54,7 +54,7 @@ const Diary = props => {
   return (
     <>
       <EmotionFilter filter={filter} changeFilter={changeFilter} />
-      <Posts
+      <Post
         list={list}
         pageNumber={Math.ceil(diaryNumber[filter] / LIMIT)}
         currentPage={currentPage}
